Код из main.js разнесен по модулям
